refactor: BIG change from Object Literals to Class
Two object literals oBoardInit and oMemoryGame are now objects.
It is easier to use and scalable this code. Easier to
understand also.
Another big change: inside these classes I use only this instead of
oMemoryGame."" that I have done before. This was a mistake!
The setTimeout function was moved inside the class. It is written
in ES6.