feat/style: add congratulations Popup/ design
Add a nice Popup Menu with great javascript code.

The html elements (for example: span-timer-s, span-hits) changed
from id to class because we can use it to the popup menu without
needing to reprogram it again.
The stars was better to create a new one.
For the stars and HTML tags as span-hits, span-miss and so on,
I have to use in this case querySelectAll instead of getElementById.
We are also looping in it. Because querySelectAll return a list
of elements.

The Design was improved: add margin left and right. Popup menu
text colors and so on.